Remove clock_gettime_GLIBC_2.17 dependency - ref #181

+++ server-plugin/Code/Systems/Blockers/RadarHackBlocker.h
@@ -84,7 +84,7 @@ private:
 
 	void* m_cvar_forcecamera;
 
-	float m_next_process;
+	double m_next_process;
 
 public:
 	RadarHackBlocker ();
@@ -99,7 +99,7 @@ private:
 
 	virtual bool GotJob () const override final;
 
-	virtual void RT_ProcessOnTick ( float const & curtime ) override final;
+	virtual void RT_ProcessOnTick (double const & curtime ) override final;
 
 	virtual void RT_ThinkPostCallback ( SourceSdk::edict_t const * const pent ) override final;
